- Move over several functions from Disassembler to Binary because it made more sense - Updated test cases to reflect changes
