+++ test/test_ghidra.py
@@ -3,7 +3,7 @@ import os
 import tempfile
 from urllib.request import urlopen
 
-from binocular import Binary, Ghidra
+from binocular import Ghidra
 
 
 def test_install_release_12():
@@ -76,56 +76,20 @@ def test_disassm(make):
             assert f0 == f1
 
 
-def test_binary(make):
-    assert Ghidra.is_installed()
-    with Ghidra("example") as g:
-        g.analyze()
-        b = g.binary
-
-        borm = b.orm()
-        assert b.architecture == borm.architecture
-        assert b.endianness == borm.endianness
-        assert b.bitness == borm.bitness
-        assert b.entrypoint == borm.entrypoint
-        assert b.os == borm.os
-        assert b.sha256 == borm.sha256
-
-        b1 = Binary.from_orm(borm)
-        assert b1.architecture == borm.architecture
-        assert b1.endianness == borm.endianness
-        assert b1.bitness == borm.bitness
-        assert b1.entrypoint == borm.entrypoint
-        assert b1.os == borm.os
-        assert b1.sha256 == borm.sha256
-
-        assert b1.architecture == b.architecture
-        assert b1.endianness == b.endianness
-        assert b1.bitness == b.bitness
-        assert b1.entrypoint == b.entrypoint
-        assert b1.os == b.os
-        assert b1.sha256 == b.sha256
-
-
 def test_function(make):
     assert Ghidra.is_installed()
     with Ghidra("example") as g:
         g.analyze()
-        f = g.function_sym("foo")
+        binary = g.binary
+        f = binary.function_sym("foo")
 
-        form = f.orm()
-        assert f.architecture == form.architecture
-        assert f.endianness == form.endianness
-        assert f.bitness == form.bitness
-        assert f.return_type == form.return_type
-        assert ", ".join([str(x) for x in f.argv]) == form.argv
-
-        f = g.function_sym("main")
+        f = binary.function_sym("main")
         # print(f.calls_addrs)
-        assert g.function_sym("foo") in [x for x in f.calls]
-        assert g.function_sym("fib") in [x for x in f.calls]
+        assert binary.function_sym("foo") in [x for x in f.calls]
+        assert binary.function_sym("fib") in [x for x in f.calls]
 
         # Recursive, so itself should be a caller and calls
-        f = g.function_sym("fib")
+        f = binary.function_sym("fib")
         assert f in [x for x in f.callers]
         assert f in [x for x in f.calls]
